requirements.txt
aioconsole==0.2.1
autobahn==20.2.2
boto3==1.11.5
botocore==1.14.5
cffi==1.14.0
click==7.1.2
cryptography==2.8
dataclasses==0.8
docutils==0.15.2
jmespath==0.9.4
joblib==0.16.0
nltk==3.5
numpy==1.18.1
pycparser==2.19
python-dateutil==2.8.1
python-decouple==3.3
regex==2020.1.8
s3transfer==0.3.1
six==1.14.0
torch==1.8.1
tqdm==4.48.0
txaio==20.1.1
typing-extensions==3.7.4.3
urllib3==1.25.7
